Maximum data is a Float (allowing sub-integer increments)
Add EditFloatPreference class
Change preference layout (new class, extra params as Floats not Longs)
Change GetNetworkStats

+++ app/src/main/java/com/example/dever/nonUglyDataUsageWidget/GetNetworkStats.kt
@@ -33,10 +33,10 @@ class GetNetworkStats(private val context: Context, private var interval: Networ
         }
 
     // TODO remove coupling from here direct to PreferenceManager
-    var maxData: Long = 0L
-        get() = PreferenceManager.getDefaultSharedPreferences(context).getLong(
+    var maxData: Float = 0F
+        get() = PreferenceManager.getDefaultSharedPreferences(context).getFloat(
                 context.resources.getString(R.string.prefs_key_maxdata),
-                context.resources.getInteger(R.integer.default_maxdata).toLong().asBytes()
+                context.resources.getInteger(R.integer.default_maxdata).toFloat().asBytes()
         )
 }
